application,cast: move message handling logic to application code.
We used to rely on the 'cast.Connection' to handle setting request ids
and waiting for those requests to finish before sending back the
results. This led to a bug where the load command would hang forever,
because we weren't able to run the media finish clean up code correctly.

Now all the request id and message handling logic is moved to the
application, it should now be easier to add general logic for all
received messages.