application: can now load http urls
The load command used to assume that the argument passed was always a
local filename and would live stream that to the cast device. When
loading a content url now, the argument passed through is now checking
for 'http://' or 'https://' to determine if the argument is a url. I am
unsure if there is need to support other schemes?

Fixed a bug with the 'next' and 'previous' commands. They were always
timing out because the cast device doesn't appear to respond to those
requests. Changed these to use 'send' rather than 'sendAndWait'.