all: cast ip address to use to be specified via flags.
Add an 'addr' command line flag to accept the address of the chromecast
device to use. This may be required if the multicast DNS doesn't work
because of networking setup. This will allow the user to specify the ip
address of the device to use. This does require the user to know the ip
address, but that can be found trivially.

Fix panic when no connection available.

Clean up error messages.

Updates: #24