application,cmd: added callback for receieved events
Added a callback for anytime a chromecast sends an event. This is a
single channel that recieves the event and gives the event to any
registered callbacks.

Update the 'watch' command to only show broadcasted events by default,
can still behave as previous when '--debug' is passed.

Update the chromcast application on any 'RECEIVER_STATUS'.

Attempt to fix transcode bug.