application: use receiver volume
It looks like the chromecast reports two different volumes, one for the
receiver and one for the media being played. It looks like
'go-chromecast' sets the receiver volume but it was reporting the media
volume.

Update to store the receiver and media volumes separately, but always
report the receiver volume as that is what 'go-chromecast' is
controlling.

Fixes: #37