Use local IP from existing conn for local server
On multi-homed hosts, the correct IP address was not always being chosen
for constructing the URL to the embedded media server that is passed to
the Cast device. By using the local-side IP address of the existing
TCP connection to the Cast device, this should ensure that the IP
address used in the URL for the embedded media server is an
address that the Cast device can reach.