application: use local addr of cast connection when possible
If a network interface is not passed through, then we can use the local
address of the established cast connection. Otherwise we should get an
ip from the passed through network interface.

This was added in #50 but I forgot that go-chromecast accepts a network
interface to look for addresses on. This should fix that.