cmd/slideshow: add a slideshow command
Added a slideshow command to display images on the chromecast. These can
be shows on repeat, or not repeated. There is a duration flag for how
long an image should be displayed on the chromecast. This has all been
coded into 'go-chromecast' because I can't get the duration to work with
the chromecast.

Updated README.

Updates: #49