cmd,dns: use net.Interface when searching with mdns (#60)
Use a network interface when trying to search for cast devices using
mdns. One was already being passed in as a flag for when deciding on a
network interface to get an ip for when hosting the server.

Pass in the net.Interface to the 'application' rather than looking for an
interface at the time we need to find the ip address.

Added a dns-timeout flag so dns timeouts are configurable in-case of a
slow dns response.