dns: use channel based api to stream faster dns results
Changed the cast dns api to return a channel which is then used to
return the cast dns entry to use quicker if specifying a name, uuid or
device type since we no longer wait for all the dns entries to be
returned.

Use zeroconf rather than mdns since it provides a nicer api and allows
us to pass in a context for cancellation / timeout.

Added --first commands which will use the first cast dns entry found.

Closes: #58