Fix bug when media switches on the chromecast
There is a bug when the chromecast changes media application, and
'go-chromecast' woudl try and write to the 'mediaFinished' channel
regardless of whether it was initialised or not. This is because of the
poor design of the internals, the 'mediaFinished' is littered throughout
the codebase but is only relevant to the 'load' command. This needs some
cleaning up in a future refactor. For now just fix the bug by not
writing to the channel if it hasn't been initialised.

Updates: #70