Don't return error when we can't determine the content type
When trying to load a media file to the chromecast, we try and determine
the content type by reading the file or doing a HEAD request if it's a
URL. If we are unable to determine the content type, we should still
send it to the chromecast in-case it can work out the content type.

Since we don't transcode the URLs it doesn't make that much sense to try
and work out the content type, but will leave it as is for now.

Updates: #96