cmd/: Exit with an error if a device filter is supplied but no device found
There was some ambiguity about what would happen if a device filter
was supplied but no device with that filter was found. Previously we
would fall back to an interactive prompt if we didn't find a device with
the filter, but this doesn't really make sense. If a device filter is
specified and a device isn't found then returning an error is more
useful.

Fix a long standing bug with '--first' where if a device filter was
supplied it was ignored with '--first' and would instead return the
first device found. This was a poor API choice, and we shouldn't allow
'--first' and a device filter at the same time since they contradict
each other. Rather than return an error for now, this will just behave
as expected and only return the first cast dns entry if no device filter
was supplied. If a device filter is supplied then we will attempt to
return that.

Updates: #159